refactor: split rates positioning snapshot builder
